file renaming
https://coveord.atlassian.net/browse/KIT-4258
