+++ packages/atomic/src/components/common/item-list/table-layout.ts
@@ -7,7 +7,7 @@ import {html, TemplateResult} from 'lit';
 import {keyed} from 'lit/directives/keyed.js';
 import {map} from 'lit/directives/map.js';
 import {ref} from 'lit/directives/ref.js';
-import {tableElementTagName} from '../../search/atomic-table-result/table-element-utils';
+import {tableElementTagName} from '../../search/atomic-table-result/stencil-table-element-utils';
 import {AnyItem} from '../interface/item';
 
 interface TableColumnsProps {
