+++ packages/atomic/src/components/common/product-template/product-template-controller.ts
@@ -4,7 +4,6 @@ import {
 } from '@coveo/headless/commerce';
 import {ReactiveController, ReactiveControllerHost} from 'lit';
 import {aggregate} from '../../../utils/stencil-utils';
-import '../../atomic-component-error/atomic-component-error';
 import {ItemTarget} from '../../common/layout/display-options';
 import {getTemplateNodeType} from './product-template-common';
