+++ packages/atomic/src/components/search/atomic-table-result/stencil-table-element-utils.tsx
