+++ packages/atomic/src/components/search/atomic-table-result/table-element-utils.ts
@@ -1,0 +1,1 @@
+export const tableElementTagName = 'atomic-table-element';
